Update patch set 1
Patch Set 1:

I think the proper way to fix this would have been https://gerrit.wikimedia.org/r/c/mediawiki/core/+/279122 but it lacked consensus.

In lieu of that, I think the cleanest approach is to declare the SpecialPage_initList hook as something that needs to be mindful of potentially running in a no-session context (as it's one of a relatively small number of hooks that can be triggered from pretty much any endpoint) and needs to call User::isSafeToLoad() when appropriate.

Patch-set: 1
Reviewer: Gergő Tisza <389@e9e9afe9-4712-486d-8885-f54b72dd1951>