Add fancy block info popups from Minerva to MobileFrontend
We only used them for registered users, but MobileFrontend's
EditorOverlay code can use them for anonymous users as well.

Write some new code in EditorGateway to look up more block details
using the action API. In Minerva we did this in PHP code and put it
in mw.config(), but we can't do it for anonymous users, because that
ends up in page HTML and is cached.

Minor formatting changes:
* Parsing the block reason is done using mw.jqueryMsg() instead of
  Linker::formatComment(), which gives mostly the same results
  (most importantly: it parses links, it does not parse templates).
* Formatting expiry time and duration is done using Moment.js rather
  than the Language class methods and custom code. There are minor
  differences in formatting, but it has the same general idea
  (human-readable dates and duration without excessive precision).

Bug: T193754
Bug: T198765
Change-Id: I80564c3e6e2d0f4d7eae754f206780fb4054f254