Opening EditorOverlay should not show interface until loading has completed
When loading the editor overlay an asynchronous request is made to load the
content and check whether the user is allowed to edit.

If the user is not the editor disappears and shows the BlockDrawer. This is jarring.

Copy the logic of the LoadingOverlay and ease-in to the editor. Don't display it until
the async call has been made.

Change-Id: I78f69b7d47e9b306079c935d4717553c9860d206