Fix image overlay loading forever
ImageOverlay was not handling missing images or connection issues
causing the spinner to show indefinitely. This addresses that by hiding
the spinner and showing a user-friendly error message with a link to
retry the load. Because the details bar is unusable without the data
from api response, it also hides that until the api response has been
successfully received. Other notable changes:

- Adds a 'hasLoadError' property to Overlay which is used by
OverlayManager to determine whether or not it should try to use its
cache to show an overlay again or try to load it again (in the event of
a load error)

- If a 'route' event is triggered from router, OverlayManager will now
compare the path property of that event to the current path. If equal,
OverlayManager will replace the current overlay with a new overlay
instead of adding the new overlay to its stack.

Bug: T193172
Change-Id: I4670d44184638760ab12312b136b275a8772ece7