Update patch set 1
Patch Set 1:

(1 comment)

I'm not sure if removing the .DS_Store (OSX stuff) and .idea directory (PHPStorm working dir) is a good move. 
I agree that we should remove the .svn, we don't use it any more, there is no need to keep it.

But when it comes to `.idea`, `DS_Store`, `~*.swp` it doesn't bother us having those entries in the .gitignore dir, this is more a list of files that shouldn't be pushed, and IMHO the longer that list the better. The more entries we have there, the easier it gets to commit new changes (we don't have to worry that by mistake we add the .DS_Store file). It happend to me before, that I added `-1` because someone pushed the .DS_Store file by mistake, having that in .gitignore would save us from situations like that.

What I want to say, I prefer to add more entries instead of removing some. But thats my standing, lets see what other people have to say.

Patch-set: 1
Reviewer: Pmiazga <3945@e9e9afe9-4712-486d-8885-f54b72dd1951>