Update patch set 1
Patch Set 1: Code-Review+1

Moral support. Global gitignore is a good thing to advocate for indeed - https://help.github.com/articles/ignoring-files/#create-a-global-gitignore

As for new contributors, I expect a novice to either not notice them (given they'd likely not have a shell PS1 configured in a way that warns for untracked files), and we as maintainers should be nice and remove them in an amend before merging, and perhaps (if deemed appropiate, or recurring) take the opportunity to talk about a global gitignore.

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1