Update patch set 1
Abandoned

I'm closing this and the other related patch to avoid possible disagreements. The subject of it seems a triviality and I'm more interested in finding causes we care for in common than debating the merits of so small a thing. Thank you all for your input and consideration.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon