Textarea padding will be defined by Minerva
This drops textarea rules in MobileFrontend which
should be expressed in a consistent manner in Minerva.
This drops padding-top from 10px to 8px in the editor
overlay which will swiftly
be corrected in the follow up in I91743fdf1942a1b1b750422e973bf98fc4d106db
which should be merged shortly after this one.

Bug: T199000
Change-Id: I1ab170fc4089b0c8129a3e62ee78efad1c6709fa