Update patch set 10
Patch Set 10: Code-Review-1

To avoid a mega commit let's do the remaining fixes in several follow ups that we'll squash together.

https://gerrit.wikimedia.org/r/#/c/mediawiki/extensions/MobileFrontend/+/448362 Prepare @event documentation for jsdoc fixes the @event documentation

The remaining issue with the documentation is that methods are not being associated with their classes - for example _changeView is being identified as a global function rather than a member of the CategoryOverlay. Let's fix that in yet another follow up.
It seems adding @name CategoryOverlay#_changeView to the function addresses the problem but that seems like a pain... Do you know a better way?

Patch-set: 10
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1