Update patch set 1
Patch Set 1:

(2 comments)

I modified the code a bit and introduced couple new functions, I didn't want to make this change too big. Now the non-empty paragraph detection works without checking the `mw-empty-elt` class.

The next step would be to change the flow to something like that:
 - find the first infobox
 - if there are no noEmptyParagraphs before the infobox
 - find the first nonEmptyParagraph after the infobox.
What do you think about that? 

For now being I think that this change is sufficient to remove the regression. Later we can focus on improving the code quality and most probably the performance as currently, that code is pretty complex and heavy on dom processing.

Patch-set: 1