Add external link icon to external links in ReferenceDrawer
We were hiding background images from external links inside the
ReferenceDrawer which prevented skins such as MinervaNeue from
displaying the external link icon. This should instead be determined by
the skin. This commit removes that style as well as a padding style so
that both of these can be set by the skin (which already happens in
MinervaNeue).

Bug: T200491
Change-Id: Ia2835431cd1558b288ba29ef852611b882ddaeca