Hygiene: Remove util.isFunction (jQuery.isFunction) in favour of typeof
As of jQuery 3.3, $.isFunction is deprecated with the recommendation
to use native `typeof`. <https://api.jquery.com/jQuery.isFunction/>

Edge cases it used to account have been removed in jQuery 3, or are for
browsers we don't support, or for values we wouldn't allow in these cases,.
E.g. bugs in IE8 where native methods like HTMLElement#getAttribute had
type `null` instead of `function`.

The `mobile.startup/util` module is referenced from one other
repository (skins/MinervaNeue), but isFunction() is not.

Change-Id: I906d5558aeeea449cdf3fc63198b0edba4dba69b