Docs: Fix incorrect indexing of OO.mfExtend
Since the switch from JSDuck to JSDoc, this was mis-categorised
as a global function called "OO", instead of a static method
of a class called "OO".

In JSDuck, @member indicated what the block a member is of.
In JSDoc, @member (alias @var) indicates the name for the block.

Use @class OO instead, and JSDoc figures out the rest.

Change-Id: I355aa89252bb8f883251c31d0148f34a5dc8c0b7