When searching for lead paragraph, use context node instead of text()
The text() is a node test and it's used to select all text-node
chilren of the context node. normalize-string expects a string value.
The current implementation didn't work if the paragraph starts with
another tag, for example:
  <p><b>Arthur Conan Doyle</b><span>[...]</span>[...]</p>

Intead of using text() we could pass string(), or just ommit the
parameter, so normalize_string will use context node and extract
the text by itself.
Sadly, using the context node broke searching the lead paragraph,
as the Coordinates span node became a valid lead paragraph, addin
another check in the getLeadParagraph() xpath query would make the
code much more complex. That would also break the DRY rule, as
check for span[@id=coordinates] has to be applied both in
`hasNoNonEmptyPrecedingParagraphs()` and in `getLeadParagraph()`.
I decided to not use `hasNoNonEmptyPrecedingParagraphs()` when
checking the infobox placement, instead code checks if the
found `$leadParagraph` is not a previous sibling of `$infobox`,
as this is the only case when we want to move the lead paragraph.

Changes:
 - instead of checking nonEmptyParagrapsh before $infobox, let's
 check that the $leadParagraph is not a $infobox previous sibling
 - removed text() call in xpath query
 - improved unit tests (better paragraph definition to include
 other tags, not only a text)
 - extracted isNonLeadParagraph() function as that code
 is used both in `hasNoNonEmptyPrecedingParagraphs` and in
 `getLeadParagraph`

Bug: T201131
Change-Id: Ifc977724b4db615f01534353e14562bb5ec6d38f