Allow users to change their mobile skin preference (take 2)
This reverts commit 54858132f362e4400a965253e78fa9da5e03273f
restoring the mobile skin preference.

The previous attempt to do this made a mistake by returning early
after normalizing the skin

Although code already existed to make this possible, the option
was not writable as the option was not registered as a valid preference
inside onUserGetDefaultOptions

This provides config that is disabled by default that allows this
to show up as a preference in Special:Preference

In future we will be able to provide this option to
editors and in meantime 3rd parties can make use of
this as requested.

Changes to original patchset
* Remove early return - this is not needed and breaks beta (and other skin options)
if MFEnableMobilePreferencesSkin has been enabled
* onUserGetDefaultOptions only registers preference if MFEnableMobilePreferencesSkin
this means that if the flag is disabled, a validation error will occur when
retrieving or saving the option e.g. new mw.Api().saveOption('mobileskin', 'minerva')
(this was what was causing the failures in T201406 despite the flag being disabled
by default)

Bug: T173527
Change-Id: Ieb241d4275adaf46abeac9a84025c825a2833806