Fix jarring white flash bug on image gallery
A white flash emanating from the image was occuring on many image
overlay loads. This was occurring  because the white background color of
the image (set by us to address transparent images) was rendering before
the image fully loads. Additionally, the spinner was moving due to the
image loading creating an unpleasant effect.

This commit hides the spinner before the image loading takes place (it
is still present for the API call). It also defers adding the white
background until the image load event has fired and adds transitions to
make this load less jarring. This seems to help in most but not all
browsers. In particular, Firefox seems to still have trouble rendering
the white background before the image in some cases which may be related
to https://bugzilla.mozilla.org/show_bug.cgi?id=626613

This, in conjunction with https://gerrit.wikimedia.org/r/448019, should
make the ImageOverlay experience less jarring.

Additional Changes:

* Makes the image text transparent. Firefox shows alt text before the
image loads which also contributed to the white flash. Making the color
transparent hides this effect. This text is also shown in the details
bar (visibly) so hiding this is assumed to not be problematic. With the
filter blur in place, this text would otherwise just be blurry.

* Adjusts the  position of the right
slider button to make it flesh with the edge of the screen. Before this
commit there was 1 em of space which, if touched, would not advance the
overlay to the next image.

* Makes .filter mixin which is now currently used by both Overlay.less
and mediaViewer.less

Bug: T197110
Change-Id: I4a0a1e9a426e929027b1ea2e12a3360213006974