Remove remaining white flicker from image overlay load
The previous attempt (commit 63639b) to resolve the white flash in image
overlay helped in most browsers but did not consistently resolve the
issue in some (iOS Safari, etc). This commit attempts to coerce the
browser in painting the loaded image before the white background color
appears by applying a background color transition delay of .15s. The CSS
class that applies the white background color is only added after the
image load event fires. It is assumed that a .15s delay will be enough
time for the browser to paint the image first and the background color
second which should help eliminate any remaining flashes from happening.

Bug: T197110
Change-Id: Icf76959079346878257fd1923274400fd1fcee66