Update patch set 11
Patch Set 11:

I don't feel this is a step backwards given we are using qunit (which is what we're using under the hood in Special:JavaScript/qunit). I do however agree it's an inconsistency, but part of this work is about showcasing why this is better and helping make a proposal to run all our unit tests for all extensions/core, so bit of a chicken egg problem here. Module loading and speed is a bit advantage of this approach (which Hashar's comment alludes to).

In this particular case, the long term plan is to move this code out of MobileFrontend in a similar way to oojs ui so that it can be consumed by Minerva and we can break the dependency between Minerva and MobileFrontend (see https://phabricator.wikimedia.org/T171000). Unfortunately, it's too risky to start with a new repo, so we'll incremently work here until we feel we can pull this out.

OOjs UI uses grunt-karma [1] to run unit tests and uses `npm test` which is what we're talking about using here so. If us not using grunt-karma is the issue, I'm happy to change that (but I figured using qunit in any form was acceptable).

Long long term, I'd like to use the outputted stubs e.g. MockMediaWiki.js etc.. to write a POC in core for running all tests in this way. There's no reason to run all Qunit tests together and all extensions/skins/core should have the option to run in isolation to avoid side effects which exist in the current approach.

A -2 is quite strong so would you be willing to expand on your comment?

[1] https://github.com/wikimedia/oojs-ui/blob/master/Gruntfile.js#L479

Patch-set: 11
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>