Update patch set 2
Patch Set 2:

As mentioned, I'm not against it.

My only suggestion (in general, not about this specifically) is that if you find that current practices in your team and by other teams cause a problem or inconvenience, to seek feedback it informally via FSG, possibly with a proposal on how to solve it, or formally via TechCom. But either way, to generally avoid too much inconsistency. These things add to confusion and make it increasingly difficult for teams to collaborate and support each other.

Patch-set: 2