Update patch set 3
Patch Set 3:

Hm, I have trouble picking which battles to fight and how to fight them. This is such a trivial little thing that I think wouldn't pursue any formalization of it, like you said, but I would appreciate your general guidance on standardization of similar but slightly more important topics.

Lately, for active repos Readers Web is on the hook for, especially Popups, Minerva, and MobileFrontend, I've been trying to keep them consistent with changes like this but I realize it's not been ideal or complete. However, it has been a very efficient way to develop ideas with little friction and both prove them and demonstrate value rather than theorizing about their effectiveness while trying to change the MediaWiki world and the opinions therein.

So, what's the best way to participate with FSG? I'm reading the wiki and I see I can raise a Phab ticket or add a topic to the Etherpad. Would I then be contacted if it's accepted for discussion or just notified of the discussion outcome?

https://www.mediawiki.org/wiki/Front-end_standards_group

Patch-set: 3
Reviewer: Niedzielski <2426@e9e9afe9-4712-486d-8885-f54b72dd1951>