Update patch set 3
Patch Set 3:

FSG is an informal group. There is a phab workboard and a weekly Hangout led by Volker currently. It's basically tagging it and/or adding it to the meeting's etherpad, we gather some opinions and see if it's trivial to move on, or if not, what we should do to get there. Getting things done :)

Jon is regular at FSG and would be able to tell you more. I've had to miss a few the last couple weeks.

Regarding ESLint, FSG did play a role in adopting it, yes. It was initially experimented with at a small scale as alternative to JSCS, primarily because our upstream connection with JSCS had been quite good (somewhat facilitated by my upstream connection with jQuery Team, which had an overlap in leadership), and when ESLint joined the foundation, and JSCS merged with ESLint, it made sense to switch. They even ported our configuration over for us! First commit to eslint-config-wikimedia is from Oleg (@markelog), who then transferred the repo to wikimedia.

As for how we adopted it, there was no formality on the pattern. We settled on something in code review. There wasn't any problem with it, and as you say now, given that there isn't a change, but it's fresh, both seem equally fine in terms of maintainability, performance and compatibility. The only problem is the third option of putting them in Grunt which is inefficient (double expansion and late exclusion, multiple eslint calls, and not discoverable by standard tooling outside Grunt).

I try to be mindful of such things early on because patterns get adopted every day, unknowingly. Because when the second instance comes around, you imply the first as a pattern. See also https://humanwhocodes.com/blog/2015/05/14/the-bunny-theory-of-code/

As for how to move forward, I don't want to go around telling everyone to do things the same way. It's your own responsibility to try and find a balance between being consistent and making it easy for others to contribute and to pick the important battles, as well as to share what you learn from experimenting. So changing this is fine. All I ask is that if it truly doesn't matter, picking what others do might be nice. And that if it does matter, it helps to have a vague intention of sharing/spreading it, and perhaps even to mention that beforehand or right after at FSG.

Patch-set: 3