Update patch set 16
Patch Set 16: -Code-Review

> I don't feel this is a step backwards given we are using qunit (which is what we're using under the hood in Special:JavaScript/qunit). I do however agree it's an inconsistency, but part of this work is about showcasing why this is better and helping make a proposal to run all our unit tests for all extensions/core, so bit of a chicken egg problem here. Module loading and speed is a bit advantage of this approach (which Hashar's comment alludes to).

Removing the tests from Special:JavaScriptTest meant that Jenkins would no longer be running them as part of the shared gate job, which would be a pretty huge regression.

> In this particular case, the long term plan is to move this code out of MobileFrontend in a similar way to oojs ui so that it can be consumed by Minerva and we can break the dependency between Minerva and MobileFrontend (see https://phabricator.wikimedia.org/T171000). Unfortunately, it's too risky to start with a new repo, so we'll incremently work here until we feel we can pull this out.

I'm not really sure why this requires switching to an entirely different way of running the tests. In MediaWiki core, we move files to includes/libs/ in preparation of librarizing, and their tests are moved to tests/phpunit/includes/libs/ as well, but are still executed by the exact same test runner, in the same process.
 
> OOjs UI uses grunt-karma [1] to run unit tests and uses `npm test` which is what we're talking about using here so. If us not using grunt-karma is the issue, I'm happy to change that (but I figured using qunit in any form was acceptable).

OOUI is a separate library, which has different standards and expectations from MediaWiki extensions.

> Long long term, I'd like to use the outputted stubs e.g. MockMediaWiki.js etc.. to write a POC in core for running all tests in this way. There's no reason to run all Qunit tests together and all extensions/skins/core should have the option to run in isolation to avoid side effects which exist in the current approach.

There is a pretty good reason - it's because that's how the modules actually run in practice. Modules are not loaded in isolation, many of them run on each page, so it reasons that they should be tested in similar environments. Maybe we need better differentiation between unit and integration tests here? In any case, this method of running all tests together was approved by TechCom (well, then ArchCom) as well: <https://www.mediawiki.org/wiki/Requests_for_comment/Extensions_continuous_integration>.

I'm not really sure what the current patchset does now, so removing my -2, since it seems that what was intended in this patch was self-merged(!) in a day in a separate patch.

Patch-set: 16
Reviewer: Legoktm <458@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0