Rename mobileFrontend global to mfModules
This is causing a lot of confusion, as MobileFrontend already
creates and modifies a global on the global mw/mediawiki object
(mw.mobileFrontend) which has a different format and contents.

So far this has confused: Stephen, Timo, myself and Nick so it feels
like we need to revisit this decision.

Given I don't think we win much by debugging (we still have the
global mw.mobileFrontend for that), I think we should consider
reintroducing this when more code is ported.. possibly when we
reconsider the future of MobileFrontend's own ModuleLoader.

Change-Id: I790c537eecda5b47e9ffb8413cc828fbeed54fe3