Update patch set 4
Patch Set 4: Code-Review+1

We chatted about this in frontend standards group and the conclusion was that .eslintignore is probably the way forward, but it should be done consistently across projects.

Volker is going to setup a task to record that outcome. When done, let's link this task to this patch and then merge.

Patch-set: 4
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1