Update patch set 16
Patch Set 16:

> Removing the tests from Special:JavaScriptTest meant that Jenkins would no longer be running them as part of the shared gate job, which would be a pretty huge regression.

Thanks for this insight. We'd figured as much, so this is why we still use JavaScriptTest. You'll thus hopefully find the current code in master to your satisfaction. Review at your leisure in case there's anything additional we've missed as I can't say with 100% confidence we addressed all your concerns.

I'd still like to challenge this notion that this is a regression (Qunit is currently being used to run integration tests), but that's a challenge that can happen another day (I've added a note to myself to write an RFC).

> I'm not really sure why this requires switching to an entirely different way of running the tests.

This, like OOUI is going to eventually be a separate library. It just happens to be packaged in MobileFrontend right now because of backwards-compatibility but we're working to change. I'm happy to talk you through our plan as I can understand as an outsider looking at one patch I can imagine it's hard to see the bigger picture inside our heads. It's documented on wiki too if you care to look there (https://www.mediawiki.org/wiki/Reading/Web/Projects/Invest_in_the_MobileFrontend_%26_MinervaNeue_frontend_architecture).

When it becomes a separate library, it won't have a MediaWiki environment so that's why we wanted to make sure the tests can run in the headless fashion. We're still using Qunit and as our new patch shows, with a little added effort we can satisfy running this in two places in the meantime to avoid confusion. We still plan to use mediawiki's qunit tests for integration when that happens. That isolation will allow us to write leaner, faster unit tests

> I'm not really sure what the current patchset does now, so removing my -2, since it seems that what was intended in this patch was self-merged(!) in a day in a separate patch.
Please do dig in to the changes we made, as we did make them on the basis of your feedback and we did work our asses off to make them run in the MediaWiki QUnit runner.

Given time constraints, we had to guess what you were thinking. I hope with our further changes in this codebase and further discussion you'll be able to understand what we'll doing here.

To be clear, patch was not self merged. For your information, Nick Ray, Stephen N and I had a one hour review session where we tested and ultimately merged the patch (given its importance). You'll thus notice 3 reviews around the same time.

This current patchset shows the diff of the changes we wanted to introduce but didn't do with your feedback. We preserve them here to invite further conversation, either on the wikitech-l thread (https://lists.wikimedia.org/pipermail/wikitech-l/2018-April/089740.html) or the
project talk page (link I posted above)!

Patch-set: 16
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>