Update patch set 18
Patch Set 18:

@Jon This last comment makes sense, and I appreciate the effort on the other commits to keep compatibility with our shared infrastructure and workflows.

But, it seems this commit now only does one thing: It explicitly removes that compat.

It doesn't add the headless testing, it change whether it does or doesn't run, it only removes the thing you just added. Was that intentional, or is this result of a rebase conflict? I'm not sure what the value would be at this point to remove only that.

Patch-set: 18
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>