Update patch set 18
Patch Set 18:

This patch is marked as WIP. We don't plan to merge it, it's just there as a reminder of what moving QUnit off Special:JavaScriptTest would look like.

When I create the RFC I'll link to this patch and we can abandon it if it is causing confusion.

Patch-set: 18
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>