Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

The code looks fine, but not sure if we should make use of mw.log.deprecate here. While there are not many gadgets based on mobile, they may exist.

(That said: creating a bug in those gadgets would be good confirmation of whether that is true)

Given MobileFrontend seems to be the only user of this method in our own extensions however seems like a good justification for not following the deprecation process here. If there were usages in Minerva, I'd insist on it.

Patch-set: 2
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1