Update patch set 11
Patch Set 11: Code-Review-1

(8 comments)

Note, inside the phab ticket we noted "the View code has dependencies on templates so these may need to be retained as QUnit tests."

If we're having trouble porting these tests, it might be worth letting the tests continue to exist as they are and raising a new task to deal with them.

It feels like we are going down a rabbit hole here.

It might therefore make sense to split this out yet again, into 2 patches - 1 that adds View.js to webpack and 2 that tries to port the test to node-qunit.

Patch-set: 11
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1