Porting PageGateway to webpack
While porting PageGateway, I struggled to also move the test to
the node-qunit folder. It has similar problems to View.js in that
it makes heavy use of sinon and jQuery.

I expect adding sinon would help port these tests as we'd be able
to stub the missing jQuery methods and would recommend doing that
first.

In the meantime, we get the benefits of this file being on webpack
and we continue to receive the test coverage as before.


Bug: T203817
Change-Id: Id3b92abfcf4093faa61b09c2c178818835758f70