Show user page creation CTA for no-JS users
There is no reason for it to be restricted to JS users. When this code
was written, all of mobile editing was JS-only, so this was just for
consistency (a0e02d6f53). But it's possible to edit without JS now.

This also ensure the link is immediately visible for JS users, rather
than flashing into view when Minerva's JS code runs.

Change-Id: Ie2fc6d43ebc03626517eec21bf4738dca05152d0