Update patch set 8
Patch Set 8: Code-Review-1

Is there a reason we can't just copy and paste this to the other repository?

If this was a pre-existing module that used to be loaded here but is now loaded elsewhere, I can understand that it is an extra step. Simple, but an extra step indeed. But, this change is explicitly creating a new unused module to only load from another repository.

Aside from the tech debt it creates, I also worry that it needlessly inflates the registry. It seems like this would not be the kind of module that should be lazy-loaded in a dedicated HTTP request. Rather, it'd something that goes along with at least one other module Minerva already has.

If this is going to be a standalone module, loaded separately and needed in situations that don't always involve at least one other always loaded at the same time, then consider my concern addressed, it can be its own module.

If not, please merge it into an existing module, in either repo.

Patch-set: 8
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1