Update patch set 6
Patch Set 6:

Using the Sinon sandbox to stub & restore the setup is pretty interesting, however, I don't think it works for jsdom. Using a simple test, lets say adding the following to the Browser `isIos()` test

    document.body.appendChild( document.createElement( 'hr' ) );
    console.log( document.body.innerHTML );  

Then adding the following to the next test `Methods are cached`:
    
    console.log( document.body.innerHTML );  

You can see that the sam DOM `<hr>` is present in both tests. 

This is probably because, although Sinon resets the value of global.window to jsdom.window, jsdom.window is an object, and javascript is "pass by reference" when it comes to objects, so values get mutated on the original jsdom.window object.

Patch-set: 6
Reviewer: Jdrewniak <2950@e9e9afe9-4712-486d-8885-f54b72dd1951>