Update patch set 1
Patch Set 1:

I agree, we definitely can remove the `$.noop` as we don't need it. Also, I think that keeping the `util.noop` function just an empty funct (not calling the jquery noop) improved the code readability. 

Previously `$.noop` or `util.noop` were explicitly used to inform other devs that this function is a no-operation, it has no side effects and most probably it's used as a "NullObject" pattern. Right now, when all no-ops are defined as `function() {}` I do not have that information anymore. This confuses me a bit, as I'm not sure if the `function() {}` is expected behaviour, is it a placeholder for something to come, or is it just something we forgot to implement.

Patch-set: 1
Reviewer: Pmiazga <3945@e9e9afe9-4712-486d-8885-f54b72dd1951>