Hygiene: enable headless View tests
No unusual changes except to undo the exception in
MobileFrontend.hooks.php and to use Hogan.js directly instead of a
rather confusing and roundabout mock.

This change just enables headless View tests in bulk with little
revision to the tests themselves. A future patchset as part of the same
task can accomplish that effort.

Bug: T204786
Change-Id: Ie452a07adf79ccf656e8d0ab8ab98d0794f87ae2