Resource modules should ignore webpack bundled file
resource-modules ensures dependencies are loaded in the appropriate
order. When using webpack this protection is guaranteed and the
tool doesn't make sense. Thus, we disable error reporting on the
output files of webpack.

The tool will continue to run on other files and will flag modules
that are required but not defined via the public API.

When we are fully migrated to webpack we can consider removing this
tooling altogether.

Bug: T206174
Change-Id: I5e4eb5eefceede794fded9605037d82f59ded87b