Fix: lint ES5 syntax and fix offender
ESLint defaults to `parserOptions.ecmaVersion` v5 and `env.es6` to
false. Apply the default ESLint configuration to all of resources/,
including resources/dist/, with ESLint config and ignores applicable to
the rest of the repo disabled. Also, fix the rogue offender in
the test-only mockMediaWiki code.

Additionally, install ESLint since a new NPM script is added, lint:es5,
which invokes the ESLint executable directly. Previously, it was an
indirect dependency.

https://eslint.org/docs/user-guide/configuring

Bug: T203976
Change-Id: Iee4c0ec303c7650764a1f105875e8a33429d84cd