Drop hook usage in Watchlist mobile via off feature flag
Mobile watchlist is not a special page feed so shouldn't pretend to
be.

While this may break existing user expectations, it would be better to
know explicitly what those are.

A feature flag is added in emergencies (e.g. fatal errors) to turn it
back on where necessary.

The feature flag is not documented in README.md as it is not meant for
public consumption.

Hopefully in future this code can be removed (promptly).

Bug: T206274
Change-Id: I2cfc51497fefecefd5c82f3415e370301add14d6