Prevent editor overlay height from being calculated incorrectly
When calculating the height of the editor overlay in EditorOvelay.js#_resizeEditor
the height is calculated when the overlay is `cols="40"` wide, as
defined in mobile.editor.overlay/content.hogan.

`cols=40` is about 448px, so scrollHeight is much longer than is should
be on wider displays.

Setting `.overlay-content` to 100% width ensures that the editor textarea is as
wide as the viewport when making this calculation.

Bug: T204835
Change-Id: I3edc990bb7ed271a4efcc21ea4be08a26063c713