Fix QUnit appendChild bug with Browser.test.js
While running `supportsAnimations() - true` and `supportsAnimations() -
false`, I was getting javascript errors relating to the
document.createElement stub we create within those tests. It turns out
that browser QUnit needs the real version of that method when calling
`assert`. This commit restores the stub to the real version prior to
calling assert and fixes the errors.

Bug: T203818
Change-Id: I82462ad288a0dfd3a9bd3c3c6697befd4ef626ac