Hygiene: forbid bad build products
Webpack build products are automated and uglified. As such, they are not
usually considered during code review but may contain a Very different
executable than what is proposed in the human readable sources.

This patch forbids build product mismatches. That is, the output of the
compilation of input sources must always match the versioned output and
produce a clean Git status.

This check is currently performed in Popups. It's not a free benefit as
we must always ensure matching inputs and outputs, but that seems like a
very good thing.

Change-Id: I643f8e1d1e4269b6578c2872f0236fb8b433643e