Update patch set 14
Patch Set 14:

I think the problem was just renaming the page "children" to "subsections" in a previous patch. Should be good now. I Changed the equality test from `strictEqual` to `propEqual` as well so that it doesn't stringify the objects before comparison (given that the order of object properties is not guaranteed).

Patch-set: 14
Reviewer: Jdrewniak <2950@e9e9afe9-4712-486d-8885-f54b72dd1951>