A simplified TalkOverlay
Now that we have the promiseView component and the
we can move our ajax call out of
the TalkBoard widget

The TalkBoard widget no longer needs to extend View,
and has no need for its template - the View class gives
everything it needs - it's just an element with a className
and two children - a p element and what will eventually become
a TopicTitleList.

The TalkOverlay now takes one child - PromisedView
The PromisedView takes care of the async nature of the
component resolving to a TalkBoard

getSections is pulled out of the TalkBoard, added to PageGateway
it is passed to the promiseView pattern

An Overlay.make utility method is added after noticing the pattern
of overlay-content appearing in many places in the codebase. This will
be utilised further in a follow up but is used for the existing talk
and language overlays

Test coverage is added for talkOverlay and talkBoard as well
as the new utility method for Overlay.

Additional CSS
* Renamed TalkOverlay.less to talk.less to aid discovery
and reflect the different components it styles
* Annotated where CSS is used.

Bug: T215370
Change-Id: I42fd7b08c4b9d92dee549d06de8a0012ea037d28