mobile.startup/rlModuleLoader: Reject promise when loading fails
Changes in 9caee5e27453516368d14c4825f40d980509376e caused this code
to always return a resolved promise, even when loading the module
failed.

When an error handler is passed to .then(), it must explicitly reject
(or throw), otherwise the new promise will be resolved (same as when
calling .catch()).

Bug: T208068
Change-Id: If70b53e4eb6ce8c274026a8a8d85b033bc5df54d